// src/generated/api/types.gen.ts
export type GameGenre = 'action' | 'adventure' | 'rpg' | 'strategy' | 'shooter' | 'simulation' | 'survival' | 'sports' | 'racing' | 'indie' | 'horror';

/**
 * Тип
 */
export type GameType = 'game' | 'dlc';

/**
 * Регион
 */
export type Region = 'ru' | 'kz' | 'by' | 'ua' | 'pl' | 'tr' | 'all_world' | 'europe' | 'asia';

/**
 * Способ получения
 */
export type DeliveryType = 'steam_key' | 'steam_gift' | 'epic_key' | 'nintendo_key' | 'xbox_key' | 'playstation_key';

export type PriceVariant = {
    /**
     * Регион
     */
    region: Region;
    /**
     * Текущая цена
     */
    price: number;
    /**
     * Старая цена
     */
    oldPrice?: number;
    /**
     * Способ получения
     */
    deliveryType: DeliveryType;
    /**
     * Издание
     */
    edition: string;
};

export type FilteredGame = {
    /**
     * Slug игры
     */
    slug: string;
    /**
     * Название игры
     */
    name: string;
    /**
     * Дата релиза
     */
    releaseDate: number;
    /**
     * Тип
     */
    type: GameType;
    /**
     * Жанры игры
     */
    genres: Array<GameGenre>;
    /**
     * Изображение игры
     */
    image: string;
    /**
     * Наименьший вариант цены
     */
    priceVariant: PriceVariant;
};